requirements.txt
openai
requests
python-dotenv
autopep8
pycodestyle
slack-bolt
slack-sdk
toml